Updated to ASP Core 2.0 Updated Project to Visual Studio 2017 Removed a few image files that were in the project root

+++ src/FableFortuneCardList/Data/Migrations/20170414060446_cards2.cs
@@ -17,7 +17,7 @@ namespace FableFortuneCardList.Data.Migrations
                 name: "Class",
                 table: "Card",
                 nullable: false,
-                defaultValue: ClassType.Neutral);
+                defaultValue: (int)ClassType.Neutral);
         }
 
         protected override void Down(MigrationBuilder migrationBuilder)
